better detect texit outputs3

+++ cmds/typst.py
@@ -97,10 +97,12 @@ class Typst(commands.Cog):
             return
     
         name = message.content.split('\n')[0]
-        print(message.content.split('\n'), name)
+
+        if len(name) < 5:
+            return
         
         if '*' == name[0] == name[1] == name[-1] == name[-2]:
-            user_id = self.bot.nickname_cache.get_user_from_nick(message.guild.id, message.content[2:-2])
+            user_id = self.bot.nickname_cache.get_user_from_nick(message.guild.id, name[2:-2])
             if self.db.get(message.guild.id, user_id):
                 await message.delete()
